+++ inst/sql/sql_server/migrations/Migration_1-v5_2_0.sql
@@ -14,8 +14,9 @@
 --    Add:
 --   - generalizability_max_sdm
 --   - generalizability_diagnostic
+--   - unblind_for_evidence_synthesis
 ALTER TABLE @database_schema.@table_prefixcm_result ADD target_estimator VARCHAR(3), one_sided_p FLOAT, calibrated_one_sided_p FLOAT;
 ALTER TABLE @database_schema.@table_prefixcm_interaction_result ADD target_estimator VARCHAR(3);
 ALTER TABLE @database_schema.@table_prefixcm_covariate_balance ADD mean_before FLOAT, mean_after FLOAT, target_std_diff FLOAT, comparator_std_diff FLOAT, target_comparator_std_diff FLOAT;
 ALTER TABLE @database_schema.@table_prefixcm_shared_covariate_balance ADD mean_before FLOAT, mean_after FLOAT, target_std_diff FLOAT, comparator_std_diff FLOAT, target_comparator_std_diff FLOAT;
-ALTER TABLE @database_schema.@table_prefixcm_diagnostics_summary ADD generalizability_max_sdm FLOAT, generalizability_diagnostic VARCHAR(20);
+ALTER TABLE @database_schema.@table_prefixcm_diagnostics_summary ADD generalizability_max_sdm FLOAT, generalizability_diagnostic VARCHAR(20), unblind_for_evidence_synthesis INT;