+++ tests/testthat/test-eunomia.R
@@ -218,6 +218,12 @@ test_that("Multiple analyses", {
   targetComparatorOutcome <- readr::read_csv(file.path(outputFolder, "export", "cm_target_comparator_outcome.csv"), show_col_types = FALSE)
   expect_true(is.numeric(targetComparatorOutcome$outcome_of_interest))
 
+  # Verify negative controls have diagnostics:
+  ncDiagnostics <- diagnosticsSummary %>%
+    inner_join(targetComparatorOutcome) %>%
+    filter(.data$outcome_of_interest == 0)
+  expect_gt(nrow(ncDiagnostics), 0)
+
   cohorts <- data.frame(
     cohortId = c(1, 2, 998, 999, 3, 4),
     cohortName = c("e1", "e2", "e3", "e4", "o1", "o2")